requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
mistralai==0.1.3
python-telegram-bot==20.7
apscheduler==3.10.4
pydantic==2.5.3
